running config2.txt:
error is when printd iterator_counter of occupied classes((It's suppose to be in the next iteration). I guess occupied classes operation should be before assign courses to available classes/ and then it will be ok. but when I change I get into infinite loop(That's mean we don't delete courses)